issue #1 (refactor-models) : switch ITopicDAO and IChapterDAO accordingly
